+++ rust/hyperlane-base/src/settings/base.rs
@@ -10,7 +10,7 @@ use hyperlane_core::{
     InterchainGasPaymasterIndexer, Mailbox, MailboxIndexer, MultisigIsm, ValidatorAnnounce, H256,
 };
 
-use crate::db::{HyperlaneDB, DB};
+use crate::db::{HyperlaneRocksDB, DB};
 use crate::{
     settings::{
         chains::{ChainConf, RawChainConf},
@@ -172,7 +172,7 @@ impl Settings {
             .build_mailbox_indexer(domain, metrics)
             .await
             .with_context(|| format!("Building mailbox indexer for {domain}"))?;
-        let hyperlane_db = HyperlaneDB::new(domain, db);
+        let hyperlane_db = HyperlaneRocksDB::new(domain, db);
         Ok(CachingMailbox::new(
             mailbox.into(),
             hyperlane_db,
@@ -191,7 +191,7 @@ impl Settings {
         let indexer = self
             .build_interchain_gas_paymaster_indexer(domain, metrics)
             .await?;
-        let hyperlane_db = HyperlaneDB::new(domain, db);
+        let hyperlane_db = HyperlaneRocksDB::new(domain, db);
         Ok(CachingInterchainGasPaymaster::new(
             interchain_gas_paymaster.into(),
             hyperlane_db,
